Refact: particle effects functionalities.
Move code for handling worker trails and tile hit effect into the
`effects` module.

Also add a limit to the number of spawners that can trigger each frame
that's tied to the buffer size allocated for the particles.

Double the size of tile hit effect buffer.

adjust particle effects parameters.